sortflow_ml_service/requirements.txt
# requirements.txt - SortFlow ML Service

fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
scikit-learn==1.5.2
Pillow==10.4.0

pydantic==2.9.2
pydantic-settings==2.6.1
